Refactor move to /experimental/temporal_memory
